fix: ManagedUserQuery hook 디렉터리 변경
